+++ apps/indexer/lib/indexer/fetcher/celo/celo_validator.ex
